Moving feedback into features
